Adding in a liveness monitor.
Any/all of our processes could hang. They are now going to go from a
strictly (?) deterministic process network to a non-deterministic
network.

Every process will take a keepalive object. They register with that
object, and it will send out periodic `ping`s. Each process declares the
amount of time it might disappear for; if the process fails to `pong`
within that timeout (in seconds), then everything quits.

This might become more graceful; for now, it is a first step towards
also listening to signals from the OS, which is necessary for responding
to systemd in a sane/appropriate way.