error handling in api module
This commit does two large things.

1. a ping/pong keepalive monitor is threaded throughout the top-level
PROCs.

2. error handling in the api module now exists.

In the case of the first, a process now exists to ping every process
every [config] seconds. The processes (mostly) reply immediately with a
`pong`. Each process has a timeout value... the tshark process, for
example, needs to have a long timeout, because it blocks on the
exec.Command() call for 45 seconds at a time.

Some processes, however, have errors to watch out for. The report_map
process is an example. It tracks how many errors it has encountered in a
given number of minutes, and if that threshold is exceeded, then it
ceases responding to pings. This way, one or more `ping` will fail to
generate a `pong`, and the process will timeout. This will cause the
entire application to fail, and (subsequently) it will allow systemd to
restart the application.

This takes a deterministic process network and turns it into a
non-deterministic process network. (That is, every process now contains
a `select` statement, and some have time.After() channels as well within
those select statements.

:shrug: