A go version of the previous script
invoke with

go run configure-pi-wifi.go

which will return the logical name of an RAlink device attached to the
machine. The first device discovered will be returned.

-descriptor=...

will let you select what you get back. By default, it is the
logicalName, which should be "wlan1."

--descriptor=mac

will give the mac address of the hardware.