Service unit, configure wifi...
This commit should completely configure wifi. To be tested is the fact
that this playbook reboots the pi if the wifi is *not* configured yet.
Will this be OK in the pi image? I don't know.

It looks for the wifi dongle, uses it to configure some things in
/etc/udev/rules.d/ and then reboots to make sure those changes take
effect. That way, on next boot, wlan1 is guaranteed to be correct. We'll
end up in a boot loop if the dongle isn't present...

Hm.

In other news, the service unit is stubbed in. Its a mess, because I
just grabbed an example as a starting point, instead of thinking through
it. The unit "works," in that I can say `service session-counter start`,
`service session-counter status` and so on. The initial test of the code
logged a lot of errors, so I wonder if there's something weird happening
when it is running as a service (vs being run by `pi` on the command
line).