working on systemd integration
golang-systemd (or whatever it is called) is not a good library.

(github builds failing)

However, I seem to have enough of the library working to be able to tell
a 'simple' service that we're alive. we can't interact with the
watchdog, it seems.

:shrug:

We're going to simply quit when our internal HTTP error count grows too
high. We'll still hope that system restarts us.