systemd, iteration
A lot of iteration on systemd. This removes the watchdog piece, because
it is not clear that it was doing anything/working.

session-counter now exits of its own accord if things go bad too many
times. then, the systemd unit will restart it. this is as opposed to the
notify system being in place.

I'd still like some kind of live monitoring, but... we'll have to get
there. for now, a gofunc running in its own thread, timing out, will
have to be our observational tool. not ideal, but there we go.