Broke out mapping code
It is now complex enough that we will wish I had done it some other way.

This revision has the anonymization for the MFG portion of the MAC
address worked in. The unit tests suggest that it is correctly tracking
the manufacturer independent of the user. That is, two RPis show up in
the table as the same mfg, but different users, and the counting is
happening correctly.

Needs to be run for an extended period of time, but looks promising.