Possible pass-by-reference fix?
We're wondering if the hash table creation in the parser was a
"pass-by-reference" error. That is, I created a hash

hash := make(...)

then appended it

results := append(results, hash)

then reused it

hash := make(...)

this may have meant that the append was not copying the data, but
appending a reference to the slice. As a result, we'd get inconsistent
results unless explicitly testing for this problem, which I wasn't.

I have added tests, and added a deepCopy() of the hash before appending.
This should guarantee that the hashmap that goes into the slice is the
real thing.