Fixing mapping error.
closes #6

I believe that doing two passes through the data takes care of this
problem. In the first pass, indicies are never reused as we cross
session boundaries. This way, unique mappings are guaranteed across
sessions (when tables are reset). As session IDs are rewritten to be
days/dates, it means we can run a second time and make sure we're
uniquely remapping within a day.

Hence, two passes (with a reset of mapping counters) on the second pass
should clean the data up appropriately.