Adding new CSP PROCs.
Adding top-level processes from a CSP-world. blackhole() consumes data,
and... delta fans out one value to many destinations. It is a PAR delta,
so they happen "in parallel," but the PROC blocks until all the reads
happen.

Also adds the type Ping{}, which is used for a channel message where we
don't want a semantic value (e.g. bool), but instead just a "hey!"
message.