Local logging of summary data
This set of changes/corrections brings the stack to a point that summary
data logs to a local SQLite database.

Changes to structs (to support the sqlx library) and a number of small
changes to address types across the libraries were required. No changes
of algorithmic substance between the previous commit and this one.