Adding a test for the TLP
There's some invasive changes here.

Every TLP now has a 'kill' channel. This allows us to shut down the TLP.
In production, this always needs to be `nil`.

In testing, we can use it to stand up a TLP network, run it, and exit
it. This is now in the session-counter_test file.

This is not yet... thoroughly tested. However, it is now possible to run
"multiple days" of testing of the full network in a... fake network.

It's surprisingly slow (on an RPi).