incorrectly clearing temporary db
The fast hack from yesterday was... fast.

When using an in-memory temporary DB, we wipe it every night. This makes
sense.

When trying to preserve data so we can understand what is going on... it
does *not* make sense to wipe the temporary DB.

The function is useful for when the db is first set up. It should not be
run on `reset`.

Now, we process data from the day, close the db from the day, and open a
new, temporary db. This will carry the filename of the current date.