singleton for logwrapper
should let us do

lw := logwrapper.NewLogger(cfg)

to config a logger that writes to a file, but in libraries do

lw := logwrapper.NewLogger(nil)

and get a logger that (if previously configured) will also write to a
file. however, if it is a first use (from some app that doesn't init the
logger), we end up with... a logger that writes to stderr.