Reworked TLP for race hazard, added image writing
Now, there's a Req/Ack pattern in-between BatchSend and WriteImages.

We write images before attempting to send to the API. This way, we make
sure we write images to disk for unreported batches. Then, we try and
send the batches, and if we're successful, they're marked as sent... and
we never write their image to disk again.