updating session id handling
This now looks at the `durations` table explicitly to get a session id.

We need to test this.

This is a hacked in first pass. What happens if there is no `durations`
table? It should return zero, but perhaps I operate on a nil pointer. (I
think I check for that...)

Does it update correctly on reset?