ProcessData "works," minor structure change
ProcessData is shuffling the data from ephemeral to the durations
database, which is straight-forward.

We need to decide whether we actually want to track the manufacturer
type. We could anonymize it (we did before), or we could ditch this
information entirely. We are not using it, and it might not be possible
to use safely/anonymously.

The minor changes in various places come from removing MfgID from the
durations struct.