Adding transaction for InsertMany (adding InsertMany)
We insert duration rows one-by-one currently. In our end-to-end test, we
sometimes have 6K+ rows to insert. This is very slow.

Transactions speed up SQLite inserts by a... big factor.

This adds InsertMany to the database interface, abstracts out the query
formation, and uses it to queue up all of the inserts at once. The
result is a massive performance improvement.