Adding profiling to the test, clearing the hash
The hash table for the durations DB... will probably be GC'd.

I'm going to explicitly walk it at the end of each day, and wipe out
everything in the DB. This will force the GC to free the data referenced
by the structure pointers held in the table.

Need to test this in production.