Update image to build properly
Previously, the Dockerfile would extend the node:18-alpine image by
adding a layer with node_modules installed.  Running `docker-compose`
would then bind-mount the local directory on top of the mode_modules
directory, install the modules in the new layer, and then run the
service.  The application was never copied to the image.

This revision copies the appilcation to the image and installs the
node_modules.  As a result, the image may be run and it'll work
as-expected.

The docker-compose.yml file still allows the bind-mounting of the
local directory to the container so one may take advantage of hot
reloads.