Un-fakified the preload data.
This changes the libraries from made-up FCFS ids (AA000x-001) to *real*
library ids.

The library IDs chosen are for real libraries, but the data in the file
does *not* correspond to the libraries indicated. We have done this so
that we can test interface elements that require lookups using real FCFS
library ids.