Merged `main`, working on tests
I've been correcting some errors in the tests and moving the branch
forward.

So far, I'm still not clear how to get a valid JWT into the system for
correctly running the api functions that require authentication. It may
be documented and I've missed it; otherwise, we need to document that
and make it part of the test infrastructure.

I believe the updates that I'm committing are "necessary" for this
branch to move to completion. Fortunately, git lets us roll back, etc.