A few more tests pass.
The gitignores might be able to be rolled back someday; ignoring
`venv`s.

README updates to reflect fixes where I found them in the README.

Migration changes are to correct permissions; the jwt_gen is a space
cleanup.

Schema reflects running the migrations.

Compose probably has the Swagger UI, which I didn't mean to include in
this branch.

`requirements.txt` is for test setup.