SQL cleanup and test improvements.
This cleanup removes some `sensor_ids`, which appeared to be a unique
table index. However, that unique table index was being requested of
clients as part of their interation with the API. These UIDs should be
internal, and not be something that the client is responsible for
knowing.

The tables.sql file is cleaned up, as much of it was auto-generated
content from the migrations. Now, serial primary keys are indicated as
SERIAL PRIMARY KEY (which is human-readable at a glance), foreign keys
are collabpsed back into table definitions as REFERENCES... and that is
most of it.

The `500-ephemeral.sh` script now does some DB modifications to make
sure the tests pass. This will have to go away in prod. It solves some
problems for automated testing during the sprint towards demo.

The unit tests pass. The result of `beat_the_heart` has become a JSON
response, so if all goes well, we get back an `{response: OK}` object.