Splitting into static and dynamic config.
The idea is to have a container that waits for the DB to be ready, and
then runs our live config.

That live config injects data into the running database system on every
container start. The static config only happens the first time we spin
up the containers. Therefore, the former should be "baseline" infra. The
dynamic/live will be where we run, say, `dbmate up` in future
iterations/migrations.