Tests, indices, removal of defaults
This commit has three things in it:

1. Updates to tests for public endpoints. It makes sure known data is
   coming back for a given query.
2. It adds two indices (per slack) to speed up binning queries. We
   recognize this could be done other ways, but for the moment, it
provides a 10-12x speedup that we need.
3. It removes defaults from the docker-compose.yaml for the backend.

The third item may break process. If you do not have the right
environment variables set, then things will not spin up correctly. The
wiki docuemtnation is up to date with this PR.

(No changes of substance occurred to the variables; the defaults have
been removed is all. This is a move in preparation for deployment. We do
not want defaults in production. They should be env vars set by the
deploy process.)